Introduced a single code base for common modules
All 4 modules of the project share common download, loading and data
tidy functions. Rather than repeated them as was the case this common
routines are now located in loadData.R and this is called in to the 4
modules using the source function.

Please refer to
https://class.coursera.org/exdata-004/forum/thread?thread_id=38 for a
thread I created on the subject.